Revert "clean up an if statement"
This reverts commit a22f45125393935e2e0a50cdfc303440cdc946d4.

The first if statement checks if busybox cache exists, and if so, does
it need to be invalidated (rm the cache file).

The second if statement checks if the cache does NOT exist, and builds
the new busybox cache.

+++ gen_compile.sh
@@ -424,7 +424,10 @@ compile_busybox() {
 		else
 			print_info 1 "busybox: >> Using cache"
 		fi
-	else
+	fi
+
+	if [ ! -f "${BUSYBOX_BINCACHE}" ]
+	then
 		cd "${TEMP}"
 		rm -rf "${BUSYBOX_DIR}" > /dev/null
 		/bin/tar -jxpf ${BUSYBOX_SRCTAR} ||
